requirements.txt
pillow
customtkinter
tk